+++ ERHMS.Test/EpiInfo/Web/ServiceTest.cs
@@ -30,12 +30,15 @@ namespace ERHMS.Test.EpiInfo.Web
         public new void OneTimeTearDown()
         {
             Settings.Default.Reset();
-            using (IDbConnection connection = new SqlConnection(ConfigurationManager.ConnectionStrings["EIWS"].ConnectionString))
+            if (view.IsWebSurvey())
             {
-                DynamicParameters parameters = new DynamicParameters();
-                parameters.Add("@SurveyId", view.WebSurveyId);
-                connection.Execute("DELETE FROM SurveyResponse WHERE SurveyId = @SurveyId", parameters);
-                connection.Execute("DELETE FROM SurveyMetaData WHERE SurveyId = @SurveyId", parameters);
+                using (IDbConnection connection = new SqlConnection(ConfigurationManager.ConnectionStrings["EIWS"].ConnectionString))
+                {
+                    DynamicParameters parameters = new DynamicParameters();
+                    parameters.Add("@SurveyId", view.WebSurveyId);
+                    connection.Execute("DELETE FROM SurveyResponse WHERE SurveyId = @SurveyId", parameters);
+                    connection.Execute("DELETE FROM SurveyMetaData WHERE SurveyId = @SurveyId", parameters);
+                }
             }
         }
